refactor(reports): extract weekly summary subsystem to reports_weekly.go
